requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4
babelfish==0.5.5
flake8==3.5.0
guessit==2.1.4
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.7.1
py==1.6.0
pyaml==17.12.1
pycodestyle==2.3.1
pyflakes==1.6.0
pytest==3.8.2
python-dateutil==2.7.3
PyYAML==3.13
rebulk==0.9.0
six==1.11.0
